@@ -15,10 +15,8 @@
 from __future__ import annotations
 
 import asyncio
-import logging.config
 
 import neo4j
-from langchain_text_splitters import CharacterTextSplitter
 from neo4j_graphrag.embeddings.openai import OpenAIEmbeddings
 from neo4j_graphrag.experimental.components.embedder import TextChunkEmbedder
 from neo4j_graphrag.experimental.components.entity_relation_extractor import (
@@ -32,38 +30,20 @@ from neo4j_graphrag.experimental.components.schema import (
     SchemaProperty,
     SchemaRelation,
 )
-from neo4j_graphrag.experimental.components.text_splitters.langchain import (
-    LangChainTextSplitterAdapter,
+from neo4j_graphrag.experimental.components.text_splitters.fixed_size_splitter import (
+    FixedSizeSplitter,
 )
 from neo4j_graphrag.experimental.pipeline import Pipeline
 from neo4j_graphrag.experimental.pipeline.pipeline import PipelineResult
-from neo4j_graphrag.llm import OpenAILLM
-
-# set log level to DEBUG for all neo4j_graphrag.* loggers
-logging.config.dictConfig(
-    {
-        "version": 1,
-        "handlers": {
-            "console": {
-                "class": "logging.StreamHandler",
-            }
-        },
-        "loggers": {
-            "root": {
-                "handlers": ["console"],
-            },
-            "neo4j_graphrag": {
-                "level": "DEBUG",
-            },
-        },
-    }
-)
+from neo4j_graphrag.llm import LLMInterface, OpenAILLM
 
 
-async def main(neo4j_driver: neo4j.Driver) -> PipelineResult:
+async def define_and_run_pipeline(
+    neo4j_driver: neo4j.AsyncDriver, llm: LLMInterface
+) -> PipelineResult:
     """This is where we define and run the KG builder pipeline, instantiating a few
     components:
-    - Text Splitter: in this example we use a text splitter from the LangChain package
+    - Text Splitter: in this example we use the fixed size text splitter
     - Chunk Embedder: to embed the chunks' text
     - Schema Builder: this component takes a list of entities, relationships and
         possible triplets as inputs, validate them and return a schema ready to use
@@ -77,23 +57,15 @@ async def main(neo4j_driver: neo4j.Driver) -> PipelineResult:
     pipe = Pipeline()
     # define the components
     pipe.add_component(
-        LangChainTextSplitterAdapter(
-            # chunk_size=50 for the sake of this demo
-            CharacterTextSplitter(chunk_size=50, chunk_overlap=10, separator=".")
-        ),
+        # chunk_size=50 for the sake of this demo
+        FixedSizeSplitter(chunk_size=4000, chunk_overlap=200),
         "splitter",
     )
     pipe.add_component(TextChunkEmbedder(embedder=OpenAIEmbeddings()), "chunk_embedder")
     pipe.add_component(SchemaBuilder(), "schema")
     pipe.add_component(
         LLMEntityRelationExtractor(
-            llm=OpenAILLM(
-                model_name="gpt-4o",
-                model_params={
-                    "max_tokens": 1000,
-                    "response_format": {"type": "json_object"},
-                },
-            ),
+            llm=llm,
             on_error=OnError.RAISE,
         ),
         "extractor",
@@ -168,8 +140,23 @@ async def main(neo4j_driver: neo4j.Driver) -> PipelineResult:
     return await pipe.run(pipe_inputs)
 
 
-if __name__ == "__main__":
-    with neo4j.GraphDatabase.driver(
+async def main() -> PipelineResult:
+    llm = OpenAILLM(
+        model_name="gpt-4o",
+        model_params={
+            "max_tokens": 1000,
+            "response_format": {"type": "json_object"},
+        },
+    )
+    driver = neo4j.AsyncGraphDatabase.driver(
         "bolt://localhost:7687", auth=("neo4j", "password")
-    ) as driver:
-        print(asyncio.run(main(driver)))
+    )
+    res = await define_and_run_pipeline(driver, llm)
+    await driver.close()
+    await llm.async_client.close()
+    return res
+
+
+if __name__ == "__main__":
+    res = asyncio.run(main())
+    print(res)
