@@ -1,6 +1,6 @@
 from neo4j import GraphDatabase
 from neo4j_graphrag.llm import OpenAILLM
-from neo4j_graphrag.retrievers.text2cypher import Text2CypherRetriever
+from neo4j_graphrag.retrievers import Text2CypherRetriever
 
 URI = "neo4j://localhost:7687"
 AUTH = ("neo4j", "password")
