@@ -4,7 +4,7 @@ import random
 import string
 
 from neo4j import GraphDatabase
-from neo4j_graphrag.embedder import Embedder
+from neo4j_graphrag.embeddings.base import Embedder
 from neo4j_graphrag.indexes import create_vector_index
 from neo4j_graphrag.retrievers import VectorRetriever
 
