@@ -15,26 +15,56 @@
 
 from __future__ import annotations
 
+import abc
 from typing import Any
 
-from neo4j_graphrag.embedder import Embedder
+from neo4j_graphrag.embeddings.base import Embedder
+
+try:
+    import openai
+except ImportError:
+    openai = None  # type: ignore
 
 
-class OpenAIEmbeddings(Embedder):
-    def __init__(self, model: str = "text-embedding-ada-002") -> None:
-        try:
-            import openai
-        except ImportError:
+class BaseOpenAIEmbeddings(Embedder, abc.ABC):
+    def __init__(self, model: str = "text-embedding-ada-002", **kwargs: Any) -> None:
+        if openai is None:
             raise ImportError(
                 "Could not import openai python client. "
                 "Please install it with `pip install openai`."
             )
-
-        self.openai_model = openai.OpenAI()
         self.model = model
 
+
+class OpenAIEmbeddings(BaseOpenAIEmbeddings):
+    """
+    OpenAI embeddings class.
+    This class uses the OpenAI python client to generate embeddings for text data.
+
+    Args:
+        model (str): The name of the OpenAI embedding model to use. Defaults to "text-embedding-ada-002".
+        kwargs: All other parameters will be passed to the openai.OpenAI init.
+    """
+
+    def __init__(self, model: str = "text-embedding-ada-002", **kwargs: Any) -> None:
+        super().__init__(model, **kwargs)
+        self.openai_client = openai.OpenAI(**kwargs)
+
     def embed_query(self, text: str, **kwargs: Any) -> list[float]:
-        response = self.openai_model.embeddings.create(
+        """
+        Generate embeddings for a given query using a OpenAI text embedding model.
+
+        Args:
+            text (str): The text to generate an embedding for.
+            **kwargs (Any): Additional arguments to pass to the OpenAI embedding generation function.
+        """
+        response = self.openai_client.embeddings.create(
             input=text, model=self.model, **kwargs
         )
         return response.data[0].embedding
+
+
+class AzureOpenAIEmbeddings(OpenAIEmbeddings):
+    def __init__(self, model: str = "text-embedding-ada-002", **kwargs: Any) -> None:
+        super().__init__(model, **kwargs)
+        self.openai_client = openai.AzureOpenAI(**kwargs)
