@@ -46,7 +46,11 @@ logger = logging.getLogger(__name__)
 
 class OnError(enum.Enum):
     RAISE = "RAISE"
-    IGNORE = "CONTINUE"
+    IGNORE = "IGNORE"
+
+    @classmethod
+    def possible_values(cls) -> List[str]:
+        return [e.value for e in cls]
 
 
 def balance_curly_braces(json_string: str) -> str:
