@@ -15,7 +15,7 @@
 
 import pytest
 from neo4j import Driver
-from neo4j_graphrag.embedder import Embedder
+from neo4j_graphrag.embeddings.base import Embedder
 from neo4j_graphrag.retrievers import VectorCypherRetriever, VectorRetriever
 from neo4j_graphrag.types import RetrieverResult, RetrieverResultItem
 
